refactor: add minor add updates/cleanups
