+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/components/SideBar/index.tsx
@@ -6,7 +6,7 @@ import {
   ArchiveIcon,
   SettingsIcon,
   ProfileIcon,
-} from '../../../components/Icons';
+} from '../Icons';
 
 function Sidebar() {
   return (
