+++ web/apps/nextjs__raw_krabby_chat_api_integrations/app/(routes)/(chat)/page.tsx
@@ -2,12 +2,12 @@
 import React from 'react';
 import Link from 'next/link';
 import { SearchIcon, PlusIcon } from './components/Icons';
-import Sidebar from './[chatId]/components/SideBar';
+import Sidebar from './components/SideBar';
 import MobileBottomNav from './components/MobileBottomNav';
 import { getInitials } from './utils';
 import { useAppDispatch, useAppSelector } from '@/app/rtk-base/hooks';
 import { openModal } from '@/app/rtk-base/slices/new_chat_modal_slice';
-import NewChatModal from './components/NewChatModal';
+import NewChatModal from './components/SideBar';
 
 export default function ChatHome() {
   const dispatch = useAppDispatch();
